+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Official support and testing for Python 3.13 ([#25](https://github.com/stumpylog/tika-client/pull/25))
+- Support for setting PDF metadata ([#42](https://github.com/stumpylog/tika-client/pull/42))
+  - Initial work by @spechtx in ([#39](https://github.com/stumpylog/tika-client/pull/42))
 
 ### Changed